Move `IndexHashToOffsetDB` and `PieceOffsetToIndexDb` into their own modules, no changes except visibility modifiers
